Backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
motor           # Async MongoDB driver
python-dotenv  # Environment variables
passlib[bcrypt]  # Password hashing
python-jose   # JWT tokens
email-validator  # For email validation in Pydantic
aiosmtplib     # Async SMTP for sending emails
pandas          # For data manipulation and Excel export
openpyxl        # For Excel file generation
